Add eslint settings and reorganize files

Add files_sharing package-lock
