+++ apps/files_sharing/.eslintrc.js
@@ -1,0 +1,22 @@
+module.exports = {
+	env: {
+		browser: true,
+		es6: true
+	},
+	globals: {
+		t: true,
+		n: true,
+		OC: true,
+		OCA: true
+	},
+	extends: 'eslint:recommended',
+	parserOptions: {
+		sourceType: 'module'
+	},
+	rules: {
+		indent: ['error', 'tab'],
+		'linebreak-style': ['error', 'unix'],
+		quotes: ['error', 'single'],
+		semi: ['error', 'always']
+	}
+};
