+++ apps/files_sharing/src/collaborationresources.js
@@ -20,20 +20,24 @@
  *
  */
 
-import Vue from 'vue'
-import Vuex from 'vuex'
-import { PopoverMenu } from 'nextcloud-vue'
-import ClickOutside from 'vue-click-outside'
-import { VTooltip } from 'v-tooltip'
-import { Store } from './services/collections';
+import Vue from 'vue';
+import Vuex from 'vuex';
+import { PopoverMenu } from 'nextcloud-vue';
+import ClickOutside from 'vue-click-outside';
+import { VTooltip } from 'v-tooltip';
+import { CollectionStore as Store } from './collectionstore';
 
 Vue.prototype.t = t;
 
-Vue.component('PopoverMenu', PopoverMenu)
-Vue.directive('ClickOutside', ClickOutside)
-Vue.directive('Tooltip', VTooltip)
+Vue.component('PopoverMenu', PopoverMenu);
+Vue.directive('ClickOutside', ClickOutside);
+Vue.directive('Tooltip', VTooltip);
 Vue.use(Vuex);
 
-import View from './CollaborationView'
+import View from './views/CollaborationView';
 
-export { Vue, View, Store }
+export {
+	Vue,
+	View,
+	Store
+};
